xrpl-js/requirements.txt
fastapi
uvicorn
xrpl-py